# extension_host: Turn on parallel compilation