--- a/crates/rope/Cargo.toml
+++ b/crates/rope/Cargo.toml
@@ -27,7 +27,7 @@ env_logger.workspace = true
 gpui = { workspace = true, features = ["test-support"] }
 rand.workspace = true
 util = { workspace = true, features = ["test-support"] }
-criterion = { version = "0.5", features = ["html_reports"] }
+criterion.workspace = true
 
 [[bench]]
 name = "rope_benchmark"